Refactor into separate packages
