+++ debugmenu/src/main/java/com/steamclock/debugmenu/DebugMenu.kt
@@ -1,5 +1,9 @@
 package com.steamclock.debugmenu
 
+import com.steamclock.debugmenu.display.DebugMenuDisplay
+import com.steamclock.debugmenu.persistence.DebugMenuPersistence
+import com.steamclock.debugmenu.persistence.readValue
+import com.steamclock.debugmenu.persistence.writeValue
 import java.util.*
 
 class DebugMenu private constructor(private val code: String = UUID.randomUUID().toString()) {
