+++ debugmenu/src/main/java/com/steamclock/debugmenu/DebugMenuState.kt
@@ -1,5 +1,10 @@
 package com.steamclock.debugmenu
 
+import com.steamclock.debugmenu.display.DebugMenuDisplay
+import com.steamclock.debugmenu.display.LogDisplay
+import com.steamclock.debugmenu.persistence.DebugMenuPersistence
+import com.steamclock.debugmenu.persistence.InMemoryPersistence
+
 data class DebugMenuState(val title: String,
                           val options: Map<String, List<DebugOption>> = mapOf(),
                           val persistence: DebugMenuPersistence = InMemoryPersistence(),
