+++ debugmenu/src/main/java/com/steamclock/debugmenu/display/DebugMenuDisplay.kt
@@ -1,4 +1,6 @@
-package com.steamclock.debugmenu
+package com.steamclock.debugmenu.display
+
+import com.steamclock.debugmenu.DebugOption
 
 /**
  * debugmenu
