+++ debugmenu/src/main/java/com/steamclock/debugmenu/persistence/InMemoryPersistence.kt
@@ -1,4 +1,4 @@
-package com.steamclock.debugmenu
+package com.steamclock.debugmenu.persistence
 
 import kotlinx.coroutines.flow.*
